+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
-## [0.1.0] - 2025-03-18
+## [1.0.0] - 2025-03-18
 
-- Initial release
+### Added
+- Initial release of `Romanizer` gem.
+- Added `to_roman` method to convert integers to Roman numerals.
+- Added `to_integer` method to convert Roman numerals to integers.
+- Basic test coverage for both methods.